detcore_model

Module time

Structs

DetTime

A determinstic notion of time, measuring progress of one thread.

GlobalTime

Deterministic global time, combining local times.

LogicalTime

Represents an absolute point in time in nanoseconds. Parts of this API are largely inspired by std::time::Duration. This could go to 128 bits if we need more than ~585 years of nanosecond precision.

Constants

NANOS_PER_NONDET_INSTR

Virtual nanoseconds elapsed per nondeterministic instruction other than system calls.

NANOS_PER_RCB

Virtual nanoseconds elapsed per Retired Conditional Branch.

NANOS_PER_SCHED

Virtual nanoseconds elapsed per step of the scheduler.

NANOS_PER_SYSCALL

Virtual nanoseconds elapsed per system call (uniform for now).

Type Aliases

LogicalDuration

Microseconds

The same basic type alias as nanoseconds. Just for clarity/readability.